kindle-pdf-convertor/requirements.txt
pyautogui
pillow
img2pdf